First commit
This is the monogame version of Virulent, since windows 8 can't run XNA.
Which will make it a pain to compile the Content project. For now, I
just use the binaries from the the current version of Virulent. Which at
the time of writing this was (Build 49) hash
b89be896d8f8496371f3371ed3885dc0b3716cb3