Better jump control
Still has a problem. Holding jump to charge means delaying the actual
jump. This delay is a good thing for a short duration, giving a feeling
of weight behind the jump. But the current duration is too long. Making
the charge time (hence the delay) any shorter would make "short
distance" jumps effectively impossible to do normally. Holding the space
bar for 1 or 2 frames is impossible to do consistantly. Currently at 10
frames for a "max charge", but it feels sluggish to jump because that
delay is too long.

Solution: Player has to be able to control jump strength by letting go
of space sooner or later, AFTER the jump has already occurred. Jump
trajectory has to be altered mid-jump.